Move files around for better directory structure
